update: queues and bookings endpoint
