+++ app/controllers/api/index.js
@@ -1,13 +1,15 @@
 const application = require("./application");
 const patients = require("./patients");
 const admins = require("./admins");
-const queue = require("./queue");
+const queues = require("./queues");
 const examinations = require("./examinations");
+const bookings = require("./bookings");
 
 module.exports = {
   application,
   admins,
   patients,
-  queue,
-  examinations
+  queues,
+  examinations,
+  bookings
 };
