+++ app/controllers/api/patients.js
@@ -1,11 +1,9 @@
 const patientServices = require('../../services/patients');
-const { Doctors } = require('../../models')
 const {
   checkPassword,
   createToken,
   hashPassword
 } = require('../../plugin');
-const queue = require('./queue');
 
 module.exports = {
   async register(req, res) {
@@ -20,7 +18,6 @@ module.exports = {
         dateOfBirth: req.body.dateOfBirth,
         address: req.body.address,
         gender: req.body.gender,
-        image: null,
         BPJS: req.body.BPJS || null,
         NIK: req.body.NIK,
         phoneNumber: req.body.phoneNumber,
@@ -33,7 +30,6 @@ module.exports = {
         dateOfBirth: patient.dateOfBirth,
         address: patient.address,
         gender: patient.gender,
-        image: patient.image,
         BPJS: patient.BPJS,
         NIK: patient.NIK,
         phoneNumber: patient.phoneNumber,
@@ -110,7 +106,6 @@ module.exports = {
         dateOfBirth,
         address,
         gender,
-        image,
         NIK,
         BPJS,
         phoneNumber
@@ -132,7 +127,6 @@ module.exports = {
         dateOfBirth,
         address,
         gender,
-        image,
         NIK,
         BPJS,
         phoneNumber,
